Added example opmode, motion profiling, and stateflows filter motor ccommands.

+++ electric-mayhem/src/main/java/edu/ncssm/ftc/electric_mayhem/fsm/sensors/JoystickValues.kt
@@ -1,3 +1,3 @@
 package sensors
 
-data class JoystickValues(val x : Double, val y : Double)
+data class JoystickValues(val x : Float, val y : Float)
